Sprzątanie po krzyśku
Najprawdopodobniej wylogowywanie nie działa z powodu polaczenia z dwoma bazami jednocześnie, Wojtek powiedział, że zrobi dzisiaj po zajęciach commita z przeniesieniem z bazy testowej na baze w ktorej Arek zrobil logowanie.
