+++ lib/database/add_ItemType.dart
@@ -1,21 +1,25 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'globalsClasses.dart';
+import 'package:inventory_app/database/globalsClasses.dart';
 
-class AddFloor extends StatelessWidget {
+class AddItemType extends StatelessWidget {
   final String buildingId;
+  final String floorId;
+  final String roomId;
 
-  final controllerName = TextEditingController();
-
-  AddFloor({
+  AddItemType({
     Key? key,
     required this.buildingId,
+    required this.floorId,
+    required this.roomId,
   }) : super(key: key);
+
+  final controllerName = TextEditingController();
 
   @override
   Widget build(BuildContext context) => Scaffold(
         appBar: AppBar(
-          title: const Text('Add floor'),
+          title: const Text('Add item type'),
         ),
         body: ListView(
           padding: const EdgeInsets.all(16),
@@ -27,15 +31,16 @@ class AddFloor extends StatelessWidget {
                 hintText: 'Name',
               ),
             ),
+            const SizedBox(height: 24),
             const SizedBox(height: 32),
             ElevatedButton(
-              child: const Text('Create floor'),
+              child: const Text('Create new item type'),
               onPressed: () async {
-                final floor = Floor(
+                final itemType = ItemType(
                   name: controllerName.text,
                 );
 
-                createFloor(floor);
+                createItemType(itemType);
                 Navigator.pop(context);
               },
             )
@@ -43,15 +48,19 @@ class AddFloor extends StatelessWidget {
         ),
       );
 
-  Future createFloor(Floor floor) async {
-    final docFloor = FirebaseFirestore.instance
+  Future createItemType(ItemType itemType) async {
+    final docItemType = FirebaseFirestore.instance
         .collection('Building')
         .doc(buildingId)
         .collection('Floor')
+        .doc(floorId)
+        .collection('Rooms')
+        .doc(roomId)
+        .collection('ItemTypes')
         .doc();
+    itemType.id = docItemType.id;
 
-    floor.id = docFloor.id;
-    final json = floor.toJson();
-    await docFloor.set(json);
+    final json = itemType.toJson();
+    await docItemType.set(json);
   }
 }
