+++ lib/database/edit_ItemType.dart
@@ -1,18 +1,22 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'add_Floor.dart';
-import 'add_Room.dart';
-import 'list_Floors.dart';
-import 'list_Rooms.dart';
+import 'add_Item.dart';
+import 'package:inventory_app/database/list_Items.dart';
 
-class EditBuilding extends StatelessWidget {
+class EditItemType extends StatelessWidget {
   final String name;
   final String buildingId;
+  final String floorId;
+  final String roomId;
+  final String itemTypeId;
 
-  EditBuilding({
+  EditItemType({
     Key? key,
     required this.name,
     required this.buildingId,
+    required this.floorId,
+    required this.roomId,
+    required this.itemTypeId,
   }) : super(key: key);
 
   final controllerName = TextEditingController();
@@ -20,15 +24,19 @@ class EditBuilding extends StatelessWidget {
   @override
   Widget build(BuildContext context) => Scaffold(
         appBar: AppBar(
-          title: Text('Edit building: $name'),
+          title: Text('Edit item type: $name'),
         ),
         floatingActionButton: FloatingActionButton(
           child: Icon(Icons.add),
           onPressed: () {
             Navigator.of(context).push(
               MaterialPageRoute(
-                builder: (context) => AddFloor(
+                builder: (context) => AddItem(
                   buildingId: buildingId,
+                  floorId: floorId,
+                  roomId: roomId,
+                  itemTypeId: itemTypeId,
+                  nameItemType: name,
                 ),
               ),
             );
@@ -40,7 +48,7 @@ class EditBuilding extends StatelessWidget {
             Padding(
               padding: EdgeInsets.only(bottom: 16),
               child: SizedBox(
-                height: 80,
+                height: 100,
                 width: 240,
                 child: Container(
                   decoration: BoxDecoration(
@@ -54,7 +62,7 @@ class EditBuilding extends StatelessWidget {
                   child: Padding(
                     padding: EdgeInsets.all(10),
                     child: Text(
-                      'Name: $name \nID: $buildingId',
+                      'Name: $name \nID: $itemTypeId',
                       style: TextStyle(fontSize: 20),
                     ),
                   ),
@@ -70,12 +78,19 @@ class EditBuilding extends StatelessWidget {
             ),
             const SizedBox(height: 24),
             ElevatedButton(
-              child: Text('Update'),
+              child: Text('Update Item Type'),
               onPressed: () async {
-                final docBuilding = FirebaseFirestore.instance
+                final docUser = FirebaseFirestore.instance
                     .collection('Building')
-                    .doc('$buildingId');
-                docBuilding.update({
+                    .doc(buildingId)
+                    .collection('Floor')
+                    .doc(floorId)
+                    .collection('Rooms')
+                    .doc(roomId)
+                    .collection('ItemTypes')
+                    .doc(itemTypeId);
+                // Update user
+                docUser.update({
                   'name':
                       (controllerName.text == '' ? name : controllerName.text),
                 });
@@ -83,7 +98,13 @@ class EditBuilding extends StatelessWidget {
               },
             ),
             const SizedBox(height: 24),
-            ListFloors(buildingId: buildingId),
+            DisplayItems(
+              buildingId: buildingId,
+              floorId: floorId,
+              roomId: roomId,
+              itemTypeId: itemTypeId,
+              itemTypeName: name,
+            ),
           ],
         ),
       );
