+++ lib/database/firebase_options.dart
