+++ lib/database/list_Floors.dart
@@ -1,9 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'add_Floor.dart';
-import 'edit_Floor.dart';
 import 'list_Rooms.dart';
-import 'readJSON.dart';
 import 'globalsClasses.dart';
 
 class ListFloors extends StatefulWidget {
