--- lib/pages/scanner/body_cam.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:inventory_app/pages/scanner/scanner_page.dart';
-
-class Body extends StatelessWidget {
-  const Body({super.key});
-
-  @override
-  Widget build(BuildContext context){
-    Size size = MediaQuery.of(context).size;
-    return const CameraPage();}}
