--- lib/pages/scanner/scan_place_pick.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:inventory_app/prefab/homebody.dart';
-import 'package:inventory_app/components/topbodysection.dart';
-
-class WyborMiejsca extends StatefulWidget {
-  @override
-  _WyborMiejscaState createState() => _WyborMiejscaState();
-}
-
-class _WyborMiejscaState extends State<WyborMiejsca> {
-
-  @override
-  Widget build(BuildContext context) {
-    final Size rozmiar = MediaQuery .of(context) .size;
-    return Scaffold(body: ScannerHomeBody(
-        title: "Skanownanie", size: rozmiar, location: Location.center));
-  }
-
-}
